lib/earlycpio: Add a function to search for an exact filename in cpio
A signature structure is used to verify the TDX module's authenticity and
integrity.  Typically, the TDX module and the signature structure are in
two files, that share the same name except for a ".sigstruct" extension,
e.g., "libtdx.bin" and "libtdx.bin.sigstruct".

Add a function, find_cpio_file(), that returns a file with the exact match,
so, for example, both "libtdx.bin" and "libtdx.bin.sigstruct" can be
correctly found. This is used instead of the existing find_cpio_data(),
which lists multiple files matching a leading path and could find the wrong
file.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>