TDX: Use private memory for TDX
The patch is only for internal private memory test and will not be
sent to community together with private memory patchset, instead it
should be sent with TDX patchset if private memory is merged ahead.

Turn on KVM_MEM_PRIVATE for TDX VM and exit to userspace for
tdx_map_gpa for private/shared conversion.

Signed-off-by: Chao Peng <chao.p.peng@linux.intel.com>