KVM: TDX: silently discard SMI request
TDX doesn't support SMM and SMI.  There are choices to handle SMI requests
from the TDX guest or the device model(e.g. qemu).  Silently ignore the
request or return the error somehow.  This patch implements silently
ignoring SMI request.

If the TDX guest requests SMI, it's troublesome for KVM to tell it to the
guest from deep KVM logic except killing the VM.  Although it's easy to
return an error to device model (e.g. qemu) on SMI request via
ioctl(KVM_SMI), it may be a result for the TDX guest to program the virtual
IOAPIC or other virtual devices.  The error should be reported to the TDX
guest.  Also the device model needs to be enhanced to handle an error when
requesting SMI.  For simplicity, the choice is to silently ignore SMI
requests for the TDX guest.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>