KVM: TDX: Handle MCE related operation for TDX
Because TDX doesn't support MCE, For TDX report that MCE isn't supported as
VM capability, initialize a related member of vcpu as MCE unsupported on
creation and ignore requests to set up MCE.

There are four MCE related operations: For TDX, the operation is modified
as follows.
- KVM_CAP_MCE: Report as MCE isn't supported via VM capability.
- KVM_X86_GET_MCE_CAP_SUPPORTED: Return zero as no MCE bank is supported as
                                 VM operation.
- KVM_X86_SETUP_MCE: Ignore request and set that no MCE bank supported.
- KVM_X86_SET_MCE: Return -EINVAL.  The existing logic already checks the
                   number of supported MCE banks.  So no change is
                   required.

KVM_CAP_MCE and KVM_X86_GET_MCE_CAP_SUPPORTED are KVM system operation.
They are enhanced as VM operation for TDX for compatibility and The KVM
system operation of them is kept the same.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>