KVM: TDX: Ignore INIT/SINIT
TDX doesn't support INIT/SINIT and defines the different protocols to boot
application processors.  Ignore INIT and SINIT events for the TDX guest.

The other choice is to return the error to the guest or device
model. (e.g. qemu) Sending INIT/SINIT is a result of operation local APIC
or virtual IOAPIC.  It's cumbersome to tell errors on operations for local
APIC or IOAPIC.  Given that TDX guest is paravirtualized to boot AP,
silently ignore INIT/SINIT request on vcpu.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>