KVM: TDX: Create initial guest memory
Because the guest memory is protected in TDX, the creation of the initial
guest memory requires a dedicated TDX module API instead of memcpy() for the
default VM type.

Define new subcommand for KVM_MEMORY_ENCRYPT_OP, KVM_TDX_INIT_MEM_REGION.
It assigns the guest page and copies and encrypts the guest initial memory.
At the same time, optionally it extends memory measurement of the TDX
guest.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>