KVM: TDX: Finalize VM initialization
To protect the initial contents of the TDX guest, the TDX module measures the
TDX guest during the build process as SHA-384 measurement.  the TDX guest
needs to complete the measurement of TDX contents and make it ready to run.

Add a new subcommand for KVM_MEMORY_ENCRYPT_OP to finalize the measurement
and mark the TDX VM ready to run.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>