x86: mce: mask out memory encryption id bits from address reported for MCE
With memory encryption technologies(MKTME, TDX and others), memory is
protected by encryption.  The memory encryption key is encoded as id in
high bits in physical address.  The bits used for encryption key is not
counted in struct cpuinfo_x86.x86_phys_bits.  Those bits can be masked out
by the use of struct cpuinfo_x86.x86_phys_bits.

When MCE happens, memory encryption key id is also included in the reported
address.  When kernel reporting the MCE, it tries to deduce what the
address was used for.  If encryption key id bits is set in reported
address, the attempts fails as follows.  Mask out the encryption key id by
the use of boot_cpu_data.x86_phys_bits so that the kernel correctly deduce
it.

Without the patch:
  mce: [Hardware Error]: Machine check events logged
  mce: Uncorrected hardware memory error in user-access at 84003c08ffe40
  Memory failure: 0x84003c08ff: memory outside kernel control
  mce: Memory error not recovered

With the patch:
  mce: [Hardware Error]: Machine check events logged
  mce: Uncorrected hardware memory error in user-access at 84003a2aadf80
  Memory Failure: 0x3a2aad: thp split failed
  Memory failure: 0x3a2aad: recovery action for unsplit thp: Ignored
  mce: Memory error not recovered

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>